random shit + storm adjustments. not finished coming back to it later
ethereals have a chance to skip storm levels
removed 3rd storm level
storms are only permanent on typhoon

+++ SS2-Project/Assets/Starstorm2/Modules/RunBehaviors/EtherealBehavior.cs
@@ -23,7 +23,7 @@ namespace SS2.Components
         public static GameObject shrinePrefab;
         public static GameObject portalPrefab;
 
-        public float etherealsCompleted = 0;
+        public int etherealsCompleted = 0;
         public static bool teleIsEthereal = false;
 
         public bool teleUpgraded;
